# nimo-plm/api/openapi.yaml
openapi: 3.1.0
info:
  title: nimo PLM System API
  description: |
    nimo智能眼镜PLM系统API接口规范
    
    ## 认证方式
    所有接口（除/auth/*外）需要在Header中携带JWT Token：
    ```
    Authorization: Bearer <token>
    ```
    
    ## 通用响应格式
    ```json
    {
      "code": 0,
      "message": "success",
      "data": {...}
    }
    ```
    
    ## 错误码
    - 0: 成功
    - 400xx: 参数错误
    - 401xx: 认证错误
    - 403xx: 权限错误
    - 404xx: 资源不存在
    - 500xx: 服务器错误
  version: 1.0.0
  contact:
    name: nimo PLM Team
    email: plm@bitfantasy.com
servers:
  - url: http://localhost:8001
    description: 开发环境
  - url: https://plm-api.nimo.internal
    description: 生产环境
...
paths:
...
  /api/v1/auth/feishu/login:
    get:
      tags: [Auth]
      summary: 飞书登录
      description: 重定向到飞书OAuth授权页面
      operationId: feishuLogin
      parameters:
        - name: redirect_uri
          in: query
          description: 授权后回调地址
          schema:
            type: string
      responses:
        '302':
          description: 重定向到飞书授权页面
  /api/v1/auth/feishu/callback:
    get:
      tags: [Auth]
      summary: 飞书授权回调
      description: 飞书授权完成后的回调接口
      operationId: feishuCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: 刷新Token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required:
                - refresh_token
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: 退出登录
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 退出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: 获取当前用户信息
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserDetail'

  # ============================================================
  # 用户管理接口
  # ============================================================
  /api/v1/users:
    get:
      tags: [Users]
      summary: 获取用户列表
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: keyword
          in: query
          description: 搜索关键词（姓名/工号/邮箱）
          schema:
            type: string
        - name: department_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, locked]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserListResponse'
  /api/v1/users/{id}:
    get:
      tags: [Users]
      summary: 获取用户详情
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserDetail'

  # ============================================================
  # 产品管理接口
  # ============================================================
  /api/v1/products:
    get:
      tags: [Products]
      summary: 获取产品列表
      operationId: listProducts
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: keyword
          in: query
          description: 搜索关键词（名称/编码）
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, developing, active, discontinued]
        - name: created_by
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProductListResponse'

    post:
      tags: [Products]
      summary: 创建产品
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
  /api/v1/products/{id}:
    get:
      tags: [Products]
      summary: 获取产品详情
      operationId: getProduct
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProductDetail'

    put:
      tags: [Products]
      summary: 更新产品
      operationId: updateProduct
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'

    delete:
      tags: [Products]
      summary: 删除产品
      operationId: deleteProduct
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/v1/products/{id}/release:
    post:
      tags: [Products]
      summary: 发布产品
      operationId: releaseProduct
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 发布成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
  /api/v1/product-categories:
    get:
      tags: [Products]
      summary: 获取产品类别列表
      operationId: listProductCategories
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'

  # ============================================================
  # 物料管理接口
  # ============================================================
  /api/v1/materials:
    get:
      tags: [Materials]
      summary: 获取物料列表
      operationId: listMaterials
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: keyword
          in: query
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, obsolete]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MaterialListResponse'

    post:
      tags: [Materials]
      summary: 创建物料
      operationId: createMaterial
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMaterialRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Material'
  /api/v1/materials/{id}:
    get:
      tags: [Materials]
      summary: 获取物料详情
      operationId: getMaterial
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Material'

    put:
      tags: [Materials]
      summary: 更新物料
      operationId: updateMaterial
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMaterialRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Material'
  /api/v1/material-categories:
    get:
      tags: [Materials]
      summary: 获取物料类别列表
      operationId: listMaterialCategories
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'

  # ============================================================
  # BOM管理接口
  # ============================================================
  /api/v1/products/{productId}/bom:
    get:
      tags: [BOM]
      summary: 获取产品BOM
      operationId: getProductBOM
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: query
          description: BOM版本号，不填返回最新版本
          schema:
            type: string
        - name: expand_level
          in: query
          description: 展开层级，-1表示全部展开
          schema:
            type: integer
            default: -1
        - name: include_cost
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BOMDetail'
  /api/v1/products/{productId}/bom/versions:
    get:
      tags: [BOM]
      summary: 获取产品BOM版本列表
      operationId: listBOMVersions
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/BOMVersion'
  /api/v1/products/{productId}/bom/items:
    post:
      tags: [BOM]
      summary: 添加BOM行项
      operationId: addBOMItem
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddBOMItemRequest'
      responses:
        '201':
          description: 添加成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BOMItem'
  /api/v1/products/{productId}/bom/items/{itemId}:
    put:
      tags: [BOM]
      summary: 更新BOM行项
      operationId: updateBOMItem
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBOMItemRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BOMItem'

    delete:
      tags: [BOM]
      summary: 删除BOM行项
      operationId: deleteBOMItem
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/v1/products/{productId}/bom/release:
    post:
      tags: [BOM]
      summary: 发布BOM版本
      operationId: releaseBOM
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReleaseBOMRequest'
      responses:
        '200':
          description: 发布成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BOMVersion'
  /api/v1/products/{productId}/bom/compare:
    get:
      tags: [BOM]
      summary: 对比BOM版本
      operationId: compareBOM
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
        - name: version_a
          in: query
          required: true
          schema:
            type: string
        - name: version_b
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BOMCompareResult'

  # ============================================================
  # 项目管理接口
  # ============================================================
  /api/v1/projects:
    get:
      tags: [Projects]
      summary: 获取项目列表
      operationId: listProjects
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: keyword
          in: query
          schema:
            type: string
        - name: product_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [planning, evt, dvt, pvt, mp, completed, cancelled]
        - name: owner_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProjectListResponse'

    post:
      tags: [Projects]
      summary: 创建项目
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Project'
  /api/v1/projects/{id}:
    get:
      tags: [Projects]
      summary: 获取项目详情
      operationId: getProject
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProjectDetail'

    put:
      tags: [Projects]
      summary: 更新项目
      operationId: updateProject
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Project'
...
  /api/v1/projects/{projectId}/tasks:
    get:
      tags: [Tasks]
      summary: 获取项目任务列表
      operationId: listProjectTasks
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: phase
          in: query
          schema:
            type: string
            enum: [evt, dvt, pvt, mp]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed, blocked, cancelled]
        - name: assignee_id
          in: query
          schema:
            type: string
        - name: overdue_only
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Task'

    post:
      tags: [Tasks]
      summary: 创建任务
      operationId: createTask
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Task'
...
  /api/v1/ecns/{id}:
    get:
      tags: [ECN]
      summary: 获取ECN详情
      operationId: getECN
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ECNDetail'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码
      schema:
        type: integer
        default: 1
        minimum: 1
...
  schemas:
...
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              nullable: true
...
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                access_token:
                  type: string
                refresh_token:
                  type: string
                expires_in:
                  type: integer
                  description: Token过期时间（秒）
                user:
                  $ref: '#/components/schemas/UserDetail'

    # ============================================================
    # 用户相关
    # ============================================================
...
security:
  - bearerAuth: []